Add files via upload
In the logs of the Noxim simulator, we demonstrated the simulation results' details reported by the deployed SystemC cycle-accurate-based simulator tool (Inspired from the Noxim).  The trace files demonstrate the cycles of all activated components, which are run using GTKwave or Modelsim, or Quartus tool. The text files illustrated the simulation results' details for no prediction, best, standard, and worst assumptions.